simplify pytest plugin
