+++ tests/integration/tests/conftest.py
@@ -10,11 +10,11 @@ import pytest
 from test_util import config, harness, tags, util
 from test_util.etcd import EtcdCluster
 from test_util.registry import Registry
-from pytest_subunit import *
+from pytest_subunit import SubunitTerminalReporter
 
 LOG = logging.getLogger(__name__)
 
-pytest_plugins = ("pytest_tagging",)
+pytest_plugins = ("pytest_tagging", )
 
 
 def pytest_itemcollected(item):
@@ -138,6 +138,13 @@ def snapd_preload() -> None:
     util.download_preloaded_snaps()
 
 
+def pytest_addoption(parser):
+    parser.addoption(
+        '--subunit-path', dest="subunit", default=None,
+        help="Stream subunit results to the specified file.")
+
+
+@pytest.hookimpl(trylast=True)
 def pytest_configure(config):
     config.addinivalue_line(
         "markers",
@@ -150,6 +157,15 @@ def pytest_configure(config):
         "node_count: Mark a test to specify how many instance nodes need to be created\n"
         "snap_versions: Mark a test to specify snap_versions for each node\n",
     )
+
+    if config.option.subunit:
+        # Get the standard terminal reporter plugin and replace it with ours.
+        standard_reporter = config.pluginmanager.getplugin('terminalreporter')
+        if not standard_reporter:
+            raise Exception("Unable to retrieve 'terminalreporter' pytest plugin.")
+        subunit_reporter = SubunitTerminalReporter(standard_reporter, config.option.subunit)
+        config.pluginmanager.unregister(standard_reporter)
+        config.pluginmanager.register(subunit_reporter, 'terminalreporter')
 
 
 @pytest.fixture(scope="function")
