+++ tests/integration/tox.ini
@@ -40,8 +40,8 @@ commands =
         --log-date-format "%Y-%m-%d %H:%M:%S" \
         --disable-warnings \
         --subunit-path "{toxinidir}/subunit.out" \
-        {posargs} \
-        {toxinidir}/tests
+        {toxinidir}/tests \
+        {posargs}
 commands_post =
     subunit2html "{toxinidir}/subunit.out" "{toxinidir}/subunit.html"
